Finalizado

+++ Exercicio101/Program.cs
@@ -1,7 +1,7 @@
 ﻿Console.Write("Digite um número: ");
 int numero = int.Parse(Console.ReadLine());
 
-long fatorial = 1;
+double fatorial = 1;
 
 for (int i = 1; i <= numero; i++)
 {
